refactoring format.c
