+++ srcs/builtins/env.c
@@ -92,3 +92,21 @@ int	ft_env(char **args, t_env *env)
 	}
 	return (0);
 }
+
+/* search key on env list. return an env node with key and value,
+	if found. else: NULL. free key*/
+t_envnode	*search_key(t_env *list, char *key)
+{
+	int			keylen;
+	t_envnode	*env;
+
+	keylen = ft_strlen(key);
+	env = list->head;
+	while (env != NULL)
+	{
+		if (ft_strncmp(key, env->key, keylen + 1) == 0)
+			break ;
+		env = env->next;
+	}
+	return (env);
+}
