+++ srcs/parser/format.c
@@ -1,61 +1,5 @@
 #include "../../includes/minishell.h"
 
-/* strlen while is a valid name for an enviroment variable ([a-z], [A-Z], '_',
-		[0-9] except on index 0, '?') */
-int	strlen_isname(char *str)
-{
-	int	i;
-
-	i = 1;
-	if (!str || str[0] == '\0')
-		return (0);
-	if (str[i] == '?')
-		return (++i);
-	while (str[i] && is_name(str[i]))
-		i++;
-	return (i);
-}
-
-/* returns index of the first quote found. if not found, returns -1 */
-int	strquote(char *str, int start)
-{
-	int	i;
-
-	i = start;
-	while (str && str[i] != '\0')
-	{
-		if (str[i] == '\'' || str[i] == '\"')
-			return (i);
-		i++;
-	}
-	return (-1);
-}
-
-/* takes a quoted str as parameter and returns it's len without outside quotes */
-int	strlen_unquote(char *quotestr)
-{
-	int	i;
-	int	start;
-	int	quote;
-
-	i = 0;
-	start = 0;
-	quote = 0;
-	while (quotestr && quotestr[i] != '\0')
-	{
-		if (quotestr[i] == '\'' || quotestr[i] == '"')
-		{
-			quote++;
-			start = i++;
-			while (quotestr[i] && quotestr[i] != quotestr[start])
-				i++;
-			quote++;
-		}
-		i++;
-	}
-	return (i - quote);
-}
-
 /* tira aspas externas do token->value */
 void	rm_quote(t_node *token)
 {
@@ -104,45 +48,6 @@ void	rm_quote(t_node *token)
 	unquoted[q] = '\0';
 	free(token->value);
 	token->value = unquoted;
-}
-
-/* search key on env list. return an env node with key and value,
-	if found. else: NULL. free key*/
-t_envnode	*search_key(t_env *list, char *key)
-{
-	int			keylen;
-	t_envnode	*env;
-
-	keylen = ft_strlen(key);
-	env = list->head;
-	while (env != NULL)
-	{
-		if (ft_strncmp(key, env->key, keylen + 1) == 0)
-			break ;
-		env = env->next;
-	}
-	return (env);
-}
-
-/* returns $ position if found outside simple quote. else: -1 */
-int	strdol(char *str)
-{
-	int	squote;
-	int	i;
-
-	squote = 0;
-	i = 0;
-	while (str && str[i] != '\0')
-	{
-		if (str[i] == '\'')
-			squote++;
-		else if (str[i] == '$' && squote % 2 == 0)
-			if (str[i + 1] == '?' || (is_name(str[i + 1]) && !ft_isdigit(str[i
-						+ 1])))
-				return (i);
-		i++;
-	}
-	return (-1);
 }
 
 /* join s1 and s2 into return string until n bytes and free s1*/
@@ -198,7 +103,6 @@ char	*ft_strfjoin(char *s1, char *s2)
 	else: its expanded to nothing. */
 void	expand(t_node *token, t_env *env)
 {
-	int			varlen;
 	char		*expanded;
 	t_envnode	*node;
 	char		*str;
@@ -213,12 +117,12 @@ void	expand(t_node *token, t_env *env)
 		if (dol < 0)
 			break ;
 		var = str + dol;
-		varlen = strlen_isname(var);
-		expanded = ft_strnjoin(expanded, str, (var - str)); // free on
-		node = search_key(env, strndup(var + 1, varlen - 1)); //free na key on
+		// expanded = ft_strnjoin(expanded, str, (var - str)); // free on
+		expanded = ft_substr(str, 0, (var - str));
+		node = search_key(env, strndup(var + 1, strlen_isname(var) - 1));
 		if (node)
 			expanded = ft_strfjoin(expanded, node->value); // free on
-		str += (var - str) + varlen;
+		str += (var - str) + strlen_isname(var);
 	}
 	expanded = ft_strfjoin(expanded, str);
 	free(token->value);
